feat: add GoNoto as a fallback font for other languages
I'm not sure about this; it's quite large and it'd be nice to have everything be the same font. I tried using it as the primary font, but I prefer the existing font. I think I'll leave it like this for now, but I wonder if there are any nice-looking universal fonts I can use?